requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow
Pillow==10.0.0
numpy
gunicorn==21.2.0
requests
gdown


